% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fit_c_svm_qp}
\alias{fit_c_svm_qp}
\title{C-svm for classification, qp solver}
\usage{
fit_c_svm_qp(y, X, kernel = poly_kernel, param.kernel = 1, c = 1,
  epsilon = 1e-06)
}
\arguments{
\item{y}{vector of class labels as -1 or 1}

\item{X}{matirx of input data, columns as samples}

\item{kernel}{kernel function, polynomial kernel by default}

\item{param.kernel}{kernel parameter, 1 by default}

\item{c}{cost parameter, 1 by default}

\item{epsilon}{threshold for selecting support vectors, 1e-6 by default}
}
\description{
C-svm for classification, qp solver
}

